:truck: chore(api): move Certification rescored by script event
