+++ api/src/certification/session-management/application/jobs/certification-rescoring-by-script-job-controller.js
@@ -1,6 +1,6 @@
-import CertificationRescoredByScript from '../../../../../lib/domain/events/CertificationRescoredByScript.js';
 import { eventDispatcher } from '../../../../../lib/domain/events/index.js';
 import { JobController } from '../../../../shared/application/jobs/job-controller.js';
+import CertificationRescoredByScript from '../../domain/events/CertificationRescoredByScript.js';
 import { CertificationRescoringByScriptJob } from '../../domain/models/CertificationRescoringByScriptJob.js';
 
 class CertificationRescoringByScriptJobController extends JobController {
